ddawdawdaw

+++ S18_ErudioMicroservices.NET6-IntegratingCheckout/GeekShopping/GeekShopping.Web/Models/CouponViewModel.cs
@@ -4,6 +4,6 @@
     {
         public long Id { get; set; }
         public string CouponCode { get; set; }
-        public double DiscountAmount { get; set; }
+        public decimal DiscountAmount { get; set; }
     }
 }
